requirements.txt
streamlit==1.45.1
numpy==1.24.0       
Pillow==11.2.1
tensorflow==2.17.1